Moved the code from the endpoints into a folder named Helpers where there is 3 classes that contains the code that is used in the endpoints.
