+++ Program.cs
@@ -1,5 +1,6 @@
 
 using GenerateDishesAPI.Data;
+using GenerateDishesAPI.Helpers;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Microsoft.Extensions.Configuration;
@@ -9,7 +10,7 @@ using Unsplasharp;
 
 namespace GenerateDishesAPI
 {
-	public class Program
+    public class Program
 	{
 		public static void Main(string[] args)
 		{
@@ -48,63 +49,25 @@ namespace GenerateDishesAPI
 
 			app.MapGet("ChatAi", async (OpenAIAPI api) =>
 			{
-				string query = "print the name of 10 different dishes(seperated with a comma)";
-
-				var chat = api.Chat.CreateConversation();
-
-				//Something went wrong here so its a bit different from how i did it but it still works
-				chat.Model = OpenAI_API.Models.Model.ChatGPTTurbo;
-
-				chat.RequestParameters.Temperature = 1;
-
-                chat.AppendSystemMessage("You are a food recepie database and will give out food recepies");
-				//Lägg till appends för json
-                chat.AppendUserInput(query);
-				chat.AppendExampleChatbotOutput("Beef wellington, Homemade pizza, Spaghetti carbonara, Birria tacos, Meatloaf, Fried chicken sandwich, Pulled pork, Escargot, Sushi, Mango sticky rice");
-				chat.AppendUserInput(query);
-				var answer = await chat.GetResponseFromChatbotAsync();
-
-                return answer;
-
-            });
+				return await OpenAiClient.GenerateTenDishes(api);
+			});
 
 			app.MapGet("/img", async (string imgQuery) =>
 			{
-				UnsplasharpClient client = new UnsplasharpClient("03jvn6lavoGJKCLyVN_Tw1GR654rFZbZbfVqRb6qiCE");
-				var photo = await client.SearchPhotos(imgQuery);
-
-				var url = photo.First().Urls.Regular;
-
-				return url;
+				return await UnsplashartAPIClient.ReturnImageUrl(imgQuery);
 			});
 
-			app.MapGet("/PicturesAndUrls", async (ApiClient client) =>
+            app.MapGet("/PicturesAndUrls", async (ApiClient client) =>
 			{
 				return await client.GeneratePicturesAndDishesAsync();
 			});
 
 			app.MapPost("GenerateRecipe", async (OpenAIAPI api, string dish) =>
 			{
-				string query = $"print the ingredients and cooking instructions for dish: {dish}";
-
-				var chat = api.Chat.CreateConversation();
-
-				//Something went wrong here so its a bit different from how i did it but it still works
-				chat.Model = OpenAI_API.Models.Model.ChatGPTTurbo;
-
-				chat.RequestParameters.Temperature = 1;
-
-				chat.AppendSystemMessage("You are a food recepie database and will give out food recepies");
-				//Lägg till appends för json
-				chat.AppendUserInput(query);
-				chat.AppendExampleChatbotOutput("Beef wellington, Homemade pizza, Spaghetti carbonara, Birria tacos, Meatloaf, Fried chicken sandwich, Pulled pork, Escargot, Sushi, Mango sticky rice");
-				chat.AppendUserInput(query);
-				var answer = await chat.GetResponseFromChatbotAsync();
-
-				return answer;
+				return await OpenAiClient.GenerateRecipeAndIngredientsForDish(api, dish);
 			});
 
-			app.Run();
+            app.Run();
 		}
 
 		//Namn email, id, lösen
